Print size of types

+++ 0x00-hello_world/6-size.c
@@ -10,14 +10,13 @@ int main(void)
 {	char c;
 	int i;
 	float f;
-	long int li;
-	long long int lli;
+	unsigned long int li;
+	unsigned long long int lli;
 
 	printf("Size of a Char: %lu byte(s) \n", (unsigned long)sizeof(c));
 	printf("Size of an int: %lu byte(s) \n", (unsigned long)sizeof(i));
 	printf("Size of a long int: %lu byte(s) \n", (unsigned long)sizeof(li));
-	printf("Size of a long long int: %lu byte(s) \n", (unsigned long)sizeof(lli));
+	printf("Size of a long long int: %lu byte(s) \n", (unsigned long long)sizeof(lli));
 	printf("Size of a float: %lu byte(s) \n", (unsigned long)sizeof(f));
 	return (0);
-
 }
